Add support for FIDO2 user verification (UV)
This is different from user presence, which only requires some kind of
interaction like touching the authenticator.

User verification checks that the user is not only present but also
authorized to use the authenticator, for example with a fingerprint
reader.

Unfortunately if a credential on the LUKS header requires UV there's
no way to know if it's usable with the inserted authenticator without
requiring some action from the user.

What we do is check all credencials without UV first, and if none of
them works then use the first credential with UV.

This means that we cannot enroll multiple credentials that require UV
from different authenticators because we have no way to know which one
is usable with the inserted one.

Closes https://github.com/bertogg/fido2luks/issues/3